requirements.txt
asgiref==3.8.1
Django==5.0.6
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
djangorestframework
django-storages
django-allauth
psycopg2-binary
requests
pyjwt
cryptography
langchain
langchain_openai
Pillow
langchain-community
boto3
pypdf
chromadb
djangorestframework-simplejwt
beautifulsoup4
django-cors-headers
pandas
openpyxl
twilio
google-api-python-client
unstructured
docx2txt
python-pptx
networkx
faiss-cpu
langchain_chroma

